llm-security-mini-project/requirements.txt
torch
transformers
accelerate
pyyaml
tqdm
matplotlib
